CopyLib.sh script file details processing. Added support for the libnss library (can be controlled by variables). Add the root user password (the default password is "root"). BusyBox compilation method adjustment.

+++ BusyBox/CopyLib.sh
@@ -9,9 +9,23 @@ fi
 AtomLinux_GraphicsLibrary="$(grep -i ^AtomLinux_GraphicsLibrary ../VariableSetting | cut -f2 -d'=')"
 AtomLinux_Only64Bit="$(grep -i ^AtomLinux_Only64Bit ../VariableSetting | cut -f2 -d'=')"
 AtomLinux_UsingWeston="$(grep -i ^AtomLinux_UsingWeston ../VariableSetting | cut -f2 -d'=')"
+AtomLinux_UsingLibnss="$(grep -i ^AtomLinux_UsingLibnss ../VariableSetting | cut -f2 -d'=')"
 #Load from VariableSetting file
 
-rm -rf ./MyConfig/lib/*
+#mkdir lib
+if [ -d ./MyConfig/lib ]; then
+    rm -rf ./MyConfig/lib/*
+else
+    mkdir ./MyConfig/lib
+fi
+if [ ${AtomLinux_Only64Bit} = "Yes" ]; then
+    if [ -d ./MyConfig/lib64 ]; then
+        rm -rf ./MyConfig/lib64/*
+    else
+        mkdir ./MyConfig/lib64
+    fi
+fi
+#mkdir lib
 
 if [ ${AtomLinux_Only64Bit} = "Yes" ]; then
     Arch="x86_64"
@@ -27,40 +41,40 @@ if [ ${Arch} = "x86_64" ]; then
 else
     ArrayLib[$((iIndex++))]="/lib/ld-linux.so.2"
 fi
-ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libc.so.6"
-ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libm.so.6"
-ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libdl.so.2"
-ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libpthread.so.0"
-ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libresolv.so.2"
+ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libc.so.6"
+ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libm.so.6"
+ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libdl.so.2"
+ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libpthread.so.0"
+ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libresolv.so.2"
 #Basic lib
 
 #GraphicsLibrary
 if [ ${AtomLinux_GraphicsLibrary} = "Qt" ]; then
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/librt.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libz.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libgcc_s.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libstdc++.so.6"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/librt.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libz.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libgcc_s.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libstdc++.so.6"
 elif [ ${AtomLinux_GraphicsLibrary} = "Qt5" ]; then
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/librt.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libz.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libgcc_s.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libudev.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libpcre.so.3"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libbsd.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libstdc++.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libdrm.so.2"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libpng16.so.16"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libGL.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libharfbuzz.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libGLX.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libGLdispatch.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libglib-2.0.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libfreetype.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libgraphite2.so.3"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libX11.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libxcb.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXau.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXdmcp.so.6"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/librt.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libz.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libgcc_s.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libudev.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libpcre.so.3"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libbsd.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libstdc++.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libdrm.so.2"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libpng16.so.16"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libGL.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libharfbuzz.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libGLX.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libGLdispatch.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libglib-2.0.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libfreetype.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libgraphite2.so.3"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libX11.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libxcb.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXau.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXdmcp.so.6"
 elif [ ${AtomLinux_GraphicsLibrary} = "Ncurses" ]; then
     :
 fi
@@ -68,38 +82,38 @@ fi
 
 #weston
 if [ ${AtomLinux_UsingWeston} = "Yes" ]; then
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/librt.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libudev.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libpcre.so.3"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libexpat.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libz.so.1"
-    ArrayLib[$((iIndex++))]="/lib/"${Arch}"-linux-gnu/libbsd.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libwayland-server.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libwayland-client.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libwayland-cursor.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libinput.so.10"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libpixman-1.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libxkbcommon.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libffi.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libmtdev.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libevdev.so.2"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libwacom.so.2"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libgudev-1.0.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libgobject-2.0.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libglib-2.0.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libcairo.so.2"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libjpeg.so.8"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libfontconfig.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libxcb-shm.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libxcb-render.so.0"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXrender.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXext.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libpng16.so.16"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libfreetype.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libxcb.so.1"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libX11.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXau.so.6"
-    ArrayLib[$((iIndex++))]="/usr/lib/"${Arch}"-linux-gnu/libXdmcp.so.6"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/librt.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libudev.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libpcre.so.3"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libexpat.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libz.so.1"
+    ArrayLib[$((iIndex++))]="/lib/${Arch}-linux-gnu/libbsd.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libwayland-server.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libwayland-client.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libwayland-cursor.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libinput.so.10"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libpixman-1.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libxkbcommon.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libffi.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libmtdev.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libevdev.so.2"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libwacom.so.2"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libgudev-1.0.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libgobject-2.0.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libglib-2.0.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libcairo.so.2"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libjpeg.so.8"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libfontconfig.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libxcb-shm.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libxcb-render.so.0"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXrender.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXext.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libpng16.so.16"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libfreetype.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libxcb.so.1"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libX11.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXau.so.6"
+    ArrayLib[$((iIndex++))]="/usr/lib/${Arch}-linux-gnu/libXdmcp.so.6"
 fi
 #weston
 
@@ -111,6 +125,15 @@ do
         fi
     fi
 done
+
+#libnss
+if [ ${AtomLinux_UsingLibnss} = "Yes" ]; then
+    MyCopy="/lib/${Arch}-linux-gnu/libnss_files*"
+    cp -dv ${MyCopy} ./MyConfig/lib/
+    MyCopy="/lib/${Arch}-linux-gnu/libnss_dns*"
+    cp -dv ${MyCopy} ./MyConfig/lib/
+fi
+#libnss
 
 #mv ld-linux to lib64
 if [ ${Arch} = "x86_64" ]; then
